Fedora: F12+ dracut initrd

+++ server/configs/k12linux/mkinitrd/ltsp-postinst.d
@@ -40,5 +40,10 @@ fi
 
 # Symlink vmlinuz.ltsp and initrd.ltsp and set permissions for tftp server
 ln -sf vmlinuz-$1    /boot/vmlinuz.ltsp
-ln -sf initrd-$1.img /boot/initrd.ltsp
+if [ -f /boot/vmlinuz-generic-$1.img ]; then
+  # dracut initrd
+  ln -sf initrd-generic-$1.img /boot/initrd.ltsp
+else
+  ln -sf initrd-$1.img /boot/initrd.ltsp
+fi
 chmod 644 /boot/initrd-$1.img
